+++ backend/src/neuroagent/tools/__init__.py
@@ -6,21 +6,88 @@ from neuroagent.tools.bluenaas_scs_getall import SCSGetAllTool
 from neuroagent.tools.bluenaas_scs_getone import SCSGetOneTool
 from neuroagent.tools.bluenaas_scs_plot import SCSPlotTool
 from neuroagent.tools.bluenaas_scs_post import SCSPostTool
-from neuroagent.tools.electrophys_tool import ElectrophysFeatureTool
+from neuroagent.tools.entitycore_asset_downloadone import AssetDownloadOneTool
+from neuroagent.tools.entitycore_asset_getall import AssetGetAllTool
+from neuroagent.tools.entitycore_asset_getone import AssetGetOneTool
+from neuroagent.tools.entitycore_brainatlas_getall import BrainAtlasGetAllTool
+from neuroagent.tools.entitycore_brainatlas_getone import BrainAtlasGetOneTool
+from neuroagent.tools.entitycore_brainregion_getall import BrainRegionGetAllTool
+from neuroagent.tools.entitycore_brainregion_getone import BrainRegionGetOneTool
+from neuroagent.tools.entitycore_brainregionhierarchy_getall import (
+    BrainRegionHierarchyGetAllTool,
+)
+from neuroagent.tools.entitycore_brainregionhierarchy_getone import (
+    BrainRegionHierarchyGetOneTool,
+)
+from neuroagent.tools.entitycore_contribution_getall import ContributionGetAllTool
+from neuroagent.tools.entitycore_contribution_getone import ContributionGetOneTool
+from neuroagent.tools.entitycore_electricalcellrecording_getall import (
+    ElectricalCellRecordingGetAllTool,
+)
+from neuroagent.tools.entitycore_electricalcellrecording_getone import (
+    ElectricalCellRecordingGetOneTool,
+)
+from neuroagent.tools.entitycore_emodel_getall import EModelGetAllTool
+from neuroagent.tools.entitycore_emodel_getone import EModelGetOneTool
+from neuroagent.tools.entitycore_etype_getall import EtypeGetAllTool
+from neuroagent.tools.entitycore_etype_getone import EtypeGetOneTool
+from neuroagent.tools.entitycore_experimentalboutondensity_getall import (
+    ExperimentalBoutonDensityGetAllTool,
+)
+from neuroagent.tools.entitycore_experimentalboutondensity_getone import (
+    ExperimentalBoutonDensityGetOneTool,
+)
+from neuroagent.tools.entitycore_experimentalneurondensity_getall import (
+    ExperimentalNeuronDensityGetAllTool,
+)
+from neuroagent.tools.entitycore_experimentalneurondensity_getone import (
+    ExperimentalNeuronDensityGetOneTool,
+)
+from neuroagent.tools.entitycore_experimentalsynapsesperconnection_getall import (
+    ExperimentalSynapsesPerConnectionGetAllTool,
+)
+from neuroagent.tools.entitycore_experimentalsynapsesperconnection_getone import (
+    ExperimentalSynapsesPerConnectionGetOneTool,
+)
+from neuroagent.tools.entitycore_ionchannelmodel_getall import (
+    IonChannelModelGetAllTool,
+)
+from neuroagent.tools.entitycore_ionchannelmodel_getone import (
+    IonChannelModelGetOneTool,
+)
+from neuroagent.tools.entitycore_measurementannotation_getall import (
+    MeasurementAnnotationGetAllTool,
+)
+from neuroagent.tools.entitycore_measurementannotation_getone import (
+    MeasurementAnnotationGetOneTool,
+)
+from neuroagent.tools.entitycore_mtype_getall import MtypeGetAllTool
+from neuroagent.tools.entitycore_mtype_getone import MtypeGetOneTool
+from neuroagent.tools.entitycore_organization_getall import OrganizationGetAllTool
+from neuroagent.tools.entitycore_organization_getone import OrganizationGetOneTool
+from neuroagent.tools.entitycore_person_getall import PersonGetAllTool
+from neuroagent.tools.entitycore_person_getone import PersonGetOneTool
+from neuroagent.tools.entitycore_reconstructionmorphology_getall import (
+    ReconstructionMorphologyGetAllTool,
+)
+from neuroagent.tools.entitycore_reconstructionmorphology_getone import (
+    ReconstructionMorphologyGetOneTool,
+)
+from neuroagent.tools.entitycore_species_getall import SpeciesGetAllTool
+from neuroagent.tools.entitycore_species_getone import SpeciesGetOneTool
+from neuroagent.tools.entitycore_strain_getall import StrainGetAllTool
+from neuroagent.tools.entitycore_strain_getone import StrainGetOneTool
+from neuroagent.tools.entitycore_subject_getall import SubjectGetAllTool
+from neuroagent.tools.entitycore_subject_getone import SubjectGetOneTool
 from neuroagent.tools.generate_plot import PlotGeneratorTool
-from neuroagent.tools.generate_random_plot import RandomPlotGeneratorTool
-from neuroagent.tools.get_morpho_tool import GetMorphoTool
-from neuroagent.tools.kg_morpho_features_tool import KGMorphoFeatureTool
 from neuroagent.tools.literature_search_tool import (
     LiteratureSearchTool,
     ParagraphMetadata,
 )
 from neuroagent.tools.morpho_metrics_tool import MorphoMetricsTool
-from neuroagent.tools.morphology_viewer import MorphologyViewerTool
 from neuroagent.tools.now import NowTool
 from neuroagent.tools.resolve_brain_region_tool import ResolveBrainRegionTool
 from neuroagent.tools.resolve_mtypes_tool import ResolveMtypeTool
-from neuroagent.tools.traces_tool import GetTracesTool
 from neuroagent.tools.weather import WeatherTool
 from neuroagent.tools.web_search import WebSearchTool
 
@@ -29,21 +96,56 @@ __all__ = [
     "SCSGetOneTool",
     "SCSPlotTool",
     "SCSPostTool",
-    "ElectrophysFeatureTool",
-    "GetMorphoTool",
-    "GetTracesTool",
-    "KGMorphoFeatureTool",
     "LiteratureSearchTool",
     "MEModelGetAllTool",
     "MEModelGetOneTool",
+    "ReconstructionMorphologyGetAllTool",
+    "ReconstructionMorphologyGetOneTool",
     "MorphoMetricsTool",
-    "MorphologyViewerTool",
     "NowTool",
     "ParagraphMetadata",
     "PlotGeneratorTool",
-    "RandomPlotGeneratorTool",
     "ResolveBrainRegionTool",
     "ResolveMtypeTool",
     "WeatherTool",
     "WebSearchTool",
+    "EtypeGetAllTool",
+    "EtypeGetOneTool",
+    "EModelGetAllTool",
+    "EModelGetOneTool",
+    "MtypeGetAllTool",
+    "MtypeGetOneTool",
+    "AssetGetAllTool",
+    "AssetGetOneTool",
+    "AssetDownloadOneTool",
+    "BrainAtlasGetAllTool",
+    "BrainAtlasGetOneTool",
+    "BrainRegionGetAllTool",
+    "BrainRegionGetOneTool",
+    "BrainRegionHierarchyGetAllTool",
+    "BrainRegionHierarchyGetOneTool",
+    "ElectricalCellRecordingGetAllTool",
+    "ElectricalCellRecordingGetOneTool",
+    "ExperimentalBoutonDensityGetAllTool",
+    "ExperimentalBoutonDensityGetOneTool",
+    "ExperimentalNeuronDensityGetAllTool",
+    "ExperimentalNeuronDensityGetOneTool",
+    "ExperimentalSynapsesPerConnectionGetAllTool",
+    "ExperimentalSynapsesPerConnectionGetOneTool",
+    "IonChannelModelGetAllTool",
+    "IonChannelModelGetOneTool",
+    "MeasurementAnnotationGetAllTool",
+    "MeasurementAnnotationGetOneTool",
+    "SpeciesGetAllTool",
+    "SpeciesGetOneTool",
+    "StrainGetAllTool",
+    "StrainGetOneTool",
+    "SubjectGetAllTool",
+    "SubjectGetOneTool",
+    "ContributionGetAllTool",
+    "ContributionGetOneTool",
+    "OrganizationGetAllTool",
+    "OrganizationGetOneTool",
+    "PersonGetAllTool",
+    "PersonGetOneTool",
 ]