+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Tools for (most) GET endpoints of entitycore
+
 ### Fixed
 - Fixed bug where fastapi servers were wrongly set.
+
+### Removed
+- Knowledge graph tools and utils
 
 ## [0.5.5] - 02.06.2025
 
@@ -26,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Morphometrics tool using obi-one.
 - Custom brain region resolving.
 - Turn `neuroagent` into an MCP client.
+- Tools for (most) GET endpoints of entitycore
 
 ### Fixed
 - Handle breaking change in entitycore.